feat : modify NotProd for initial user setting

+++ src/main/java/com/example/MnM/boundedContext/member/entity/Member.java
@@ -40,8 +40,8 @@ public class Member {
     @LastModifiedDate
     private LocalDateTime updatedAt; //수정일
     //여기까지 회원가입시 기본정보, 아래로는 개인정보
-    private Long height; //키
-    private Long age; //나이;
+    private Integer height; //키
+    private Integer age; //나이;
     private String locate; //지역
     private String gender; //성별
     private String mbti; //mbti
